simplify NOp cache loading loop
this has a side effect of changing the key assignment order in the case when several node operators have the same effective stake: previously, keys were assigned in the id-reverse order (i.e. the ope that were added later were preferred over those who were added earlier); now, in the case of the same effective stake keys are assigned in the id order.

related to: 44626e2